Changed the algorithm again
The change I made actually makes the code longer, but it makes it even
more efficient in the long term. The code now only checks prime numbers
that are less than or equal to the square root of the number we are
checking in the loop, because if you go over the square root then the
other factor is a prime we already checked.